performance tuned around string encoding
