+++ src/main/java/com/cloudurable/jparse/parser/JsonEventAbstractParser.java
@@ -1,0 +1,2 @@
+package com.cloudurable.jparse.parser;public class JsonEventAbstractParser {
+}
