+++ src/test/java/com/cloudurable/jparse/JsonParserFastEventsTest.java
@@ -1,0 +1,22 @@
+package com.cloudurable.jparse;
+
+import com.cloudurable.jparse.parser.JsonIndexOverlayParser;
+import com.cloudurable.jparse.source.CharSource;
+import com.cloudurable.jparse.token.TokenEventListener;
+
+public class JsonParserEventsTest extends JsonParserTest{
+    @Override
+    public JsonIndexOverlayParser jsonParser() {
+        return Json.builder().setStrict(true).setTokenEventListener(new TokenEventListener() {
+            @Override
+            public void start(int tokenId, int index, CharSource source) {
+
+            }
+
+            @Override
+            public void end(int tokenId, int index, CharSource source) {
+
+            }
+        }).build();
+    }
+}
