+++ src/test/java/com/cloudurable/jparse/JsonScannerEventsTest.java
@@ -1,0 +1,2 @@
+package com.cloudurable.jparse;public class JsonScannerEventsTest {
+}
